Delete commented out code
