+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/booking/payment/BookingPaymentPresenter.kt
@@ -1,10 +1,11 @@
-package com.karhoo.uisdk.screen.booking.booking
+package com.karhoo.uisdk.screen.booking.booking.payment
 
-import android.util.Log
 import com.karhoo.sdk.api.KarhooApi
 import com.karhoo.sdk.api.network.response.Resource
 import com.karhoo.sdk.api.service.payments.PaymentsService
+import com.karhoo.uisdk.R
 import com.karhoo.uisdk.base.BasePresenter
+import com.karhoo.uisdk.screen.booking.booking.payment.BookingPaymentMVP
 
 class BookingPaymentPresenter(view: BookingPaymentMVP.View,
                               private val paymentsService: PaymentsService = KarhooApi.paymentsService)
@@ -19,7 +20,7 @@ class BookingPaymentPresenter(view: BookingPaymentMVP.View,
         paymentsService.getPaymentProvider().execute { result ->
             when (result) {
                 is Resource.Success -> handleGetPaymentProviderSuccess(result.data.provider.id)
-                is Resource.Failure -> Log.d("PD36", "${result.error.internalMessage}")
+                is Resource.Failure -> view?.showError(R.string.something_went_wrong)
             }
         }
     }
